feat: Add language and automaton blog post
This commit adds a new blog post titled "Automates & Langages". The blog post is still in progress and will be completed later. It introduces a simplified grammar in EBNF notation for a subset of YAML and explains the construction of a pushdown automaton (PDA) for parsing YAML documents. The commit includes the initial content of the blog post.

Co-authored-by: Mark Horn <markhorn-dev@users.noreply.github.com>
Co-authored-by: Evan <53529192+BruceMcRooster@users.noreply.github.com>
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>